refactor(web): extract shared ActiveOverlayShell from the active-* overlays (#36064)
